requirements.txt
attrs==22.1.0
jsonschema==4.17.0
pyrsistent==0.18.1
pynacl==1.5.0
